Changed from private to public class
